web_app/requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
gunicorn==19.0.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.3
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywinpty==0.5.7
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==3.0.5
streamlit==0.71.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.58.0
traitlets==4.3.3
tzlocal==2.1
urllib3==1.26.3